+++ index.html
@@ -228,7 +228,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
+  <script src="./leaflet/leaflet.js" crossorigin=""></script>
  <script src="./js/main.js" type="module"></script>
- <script src="./leaflet/leaflet.js" crossorigin=""></script>
 </body>
 </html>
